backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
annotated-types==0.7.0
    # via pydantic
asgiref==3.8.1
    # via django
bcrypt==4.1.3
    # via -r requirements.in
cachetools==5.5.2
    # via google-auth
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
dj-database-url==2.3.0
    # via -r requirements.in
django==4.2.20
    # via
    #   -r requirements.in
    #   dj-database-url
    #   djangorestframework
    #   djangorestframework-simplejwt
djangorestframework==3.15.2
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
djangorestframework-simplejwt==5.3.1
    # via -r requirements.in
google-auth==2.39.0
    # via google-genai
google-genai==0.8.0
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
idna==3.10
    # via requests
packaging==24.2
    # via gunicorn
psycopg2-binary==2.9.10
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.11.3
    # via google-genai
pydantic-core==2.33.1
    # via pydantic
pyjwt==2.10.1
    # via djangorestframework-simplejwt
python-dotenv==1.1.0
    # via -r requirements.in
requests==2.32.3
    # via google-genai
rsa==4.9.1
    # via google-auth
sqlparse==0.5.3
    # via django
typing-extensions==4.13.2
    # via
    #   dj-database-url
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
urllib3==2.4.0
    # via requests
websockets==14.2
    # via google-genai
